Split gameSearch into it's own file with blueprint
